+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 4.1.0
+* **Feature Added** Button to change playback rate in player.
+* Added `setPlaybackRate` method to `YoutubePlayerController`. [Issue #48](https://github.com/sarbagyastha/youtube_player_flutter/issues/48) 
+* **Improvement** Playback behaviour synced with `AppLifeCycle` [Issue #41](https://github.com/sarbagyastha/youtube_player_flutter/issues/41) 
+
 ## 4.0.0+2
 * Finally working on iOS.
 * Fix error reports in LiveUI. [PR#44](https://github.com/sarbagyastha/youtube_player_flutter/pull/44)